+++ bigchaindb/db/utils.py
@@ -54,7 +54,7 @@ def init():
         .run(conn)
     # secondary index for payload hash
     r.db(dbname).table('bigchain')\
-        .index_create('payload_hash', r.row['block']['transactions']['transaction']['data']['hash'], multi=True)\
+        .index_create('payload_uuid', r.row['block']['transactions']['transaction']['data']['uuid'], multi=True)\
         .run(conn)
 
     # wait for rethinkdb to finish creating secondary indexes