+++ CHANGELOG.md
@@ -15,10 +15,23 @@ For reference, the possible headings are:
 * **Notes**
 
 
-## [0.4.1] - 2016-06-13
-Tag name: v0.4.1
+## [0.4.2] - 2016-06-15
+Tag name: v0.4.2
 = commit: 
 committed: 
+
+### Added
+- Report the BigchainDB version number when starting BigchainDB: [Pull Request #385](https://github.com/bigchaindb/bigchaindb/pull/385)
+
+### Changed
+- Round timestamps to a precision of one second, and replace payload hash with payload UUID in transactions: [Pull Request #384](https://github.com/bigchaindb/bigchaindb/pull/384)
+- Updated cryptoconditions API usage: [Pull Request #373](https://github.com/bigchaindb/bigchaindb/pull/373) 
+
+
+## [0.4.1] - 2016-06-13
+Tag name: v0.4.1
+= commit: 9c4aa987bcbc294b6a5c3069e6c45a7ed77a4068
+committed: June 13, 2016, 9:52 AM GMT+2
 
 ### Added
 - Revert `bigchain` deletes: [Pull Request #330](https://github.com/bigchaindb/bigchaindb/pull/330)